+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Socket.IsBound` implementation for `SocksClient`.
+
 ## [0.5.0-alpha] - 2024-09-06
 
 ### Added